pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.11",
]
